Move vertex examples into vertex folder
