+++ src/sample/vertex/gltfViewer/gltf.frag.wgsl
