+++ src/sample/vertex/normalMapping/enum.ts
