+++ src/sample/vertex/normalMapping/normalMap.wgsl
