+++ src/sample/vertex/simpleVertex/mesh.wgsl
